Сделал крад покруче

+++ server/frontend/controllers/issue_file/Delete.php
@@ -4,6 +4,9 @@ namespace frontend\controllers\issue_file;
 
 use frontend\components\RestAction;
 use common\models\IssueFileRecord;
+use common\models\IssueRecord;
+use common\models\ProjectRecord;
+use common\models\CustomerRecord;
 use Yii;
 
 class Delete extends RestAction
@@ -17,6 +20,14 @@ class Delete extends RestAction
 			return;
 		}
 		
+		$project = ProjectRecord::findOne(['id' => IssueRecord::findOne($issueFile->issue_id)->project_id]);
+		$customer = CustomerRecord::findOne(['user_id' => Yii::$app->user->identity->id]);
+
+		if ($project->customer_id !== $customer->id) {
+			Yii::$app->getResponse()->setStatusCode(403);
+			return;
+		}
+
 		if (false === $issueFile->delete()) {
 			throw new Exception('Failed to delete the Issue File for unknown reason.');
 		}
